Fix usage of RedisModule_HashGet in raw::hash_get and RedisModule_HashSet in raw::hash_set
The Rust value "0" was being used as the terminating null pointer signalling the end of
the variadic arguments. Unfortunately the type of this expression is i32, which creates
a "read-past-the-end" condition in the calling C code. For calls to RedisModule_HashGet
with a single argument this seems to usually work correctly, but trying to do the with
2 or more fields crashes with a segmentation fault.