Add infrastructure for exposing hash-multi-get functionality (#105)
This commit adds a `hash_get_multi` function in the `raw` module, and a private
`hash_mget_key` function in the `key` module. In the interests of avoiding code
duplication, the corresponding single-get functions are removed, and the public
RedisKey*.hash_get functionality is implemented via multi-get.

Public-facing RedisKey* methods for hash multi-get should be the subject of their
own commit.