Provide an IntoIterator for the multi-get results
This is both more idiomatic and more flexible than the previous `into_list`/`into_map`.
It is now possible to both immediately collect the results into the user's choice of
collection, and to perform further manipulations on the results via iterator methods
without materializing them into an intermediate collection and incurring allocation costs.

Since the resulting iterator iterates of (key,value) pairs, this comes at the small cost
of needing to explicitly `map` the results to collect them into a vector of field-values only.