Added RM_Call options
The PR added the ability to use RM_Call with different options
The available options are:
1. No writes
2. Script mode
3. Acl verification
4. OOM verification
5. Return errors as CallReply objects
6. Replicate the command to replica and AOF
7. Control the protocol version in which the replies will be returned

The call option can be give to `call_ext` function which is the same as `call`
but also accept the `CallOption` object.

In addition the PR allow to get the reply as A CallReply object that wraps the Redis
module API CallReply (RedisModuleCallReply*) and allow to avoid unneeded coppies and
reply parsing.