Added context mutex struct that allows to save static data protected by Redis GIL.
There are a wide range of usecases that require the module to save static data.
Most of the times this static data also need to be mutated by the module.
Rust, by definition, defines mutable static data as unsafe and accessing it,
even in a none mutable way, require unsafe code.

In order to avoid unsafe code, one must wrap the static data with a mutex.
The issue is that most of the times the data is only accessed when the Redis GIL
is held, wrapping the data with another mutex just to avoid unsafe code is redudent
and can even reduce performance.

The PR introduce a `ContextMutex` struct that just like a regular mutex, can wrap data
and make it safe to access the it. The `ContextMutex` can be locked by giving it a
reference to a `RedisContext` and the lock period can not outlive the given RedisContext
nor the `ContextMutex`.

`ContextMutex` can also safely pass between threads (implements `Sync` and `Send`). So it
can also be used to pass objects to other threads and use them (after acquire the Redis GIL).

**Notice**: The base assumption is that when the code has a reference to `RedisContext` the Redis GIL must be acquired. This assumption is currently not always true. We should find locations where we break this assumption and fix it gradually.