Added RM_Call options (#290)
The PR added the ability to use RM_Call with different options The available options are:
1. No writes
2. Script mode
3. Acl verification
4. OOM verification
5. Return errors as CallReply objects
6. Replicate the command to replica and AOF
7. Control the protocol version in which the replies will be returned

The call option can be give to `call_ext` function which is the same as `call` but also accept the `CallOption` object.

In addition the PR allow to get the reply as A CallReply object that wraps the Redis module API CallReply (RedisModuleCallReply*) and allow to avoid unneeded coppies and reply parsing.

### Low Level Implementation details

To properly return `CallReply` object we need to be able to define a proper lifetime for `CallReply`. The main problem is that Redis makes no distinction between root call reply and inner call reply (that is reachable from the root). The module writer need to make sure he only free the root call reply (and never hold some inner call reply for a longer time then its root).

To expose those limitation we implement 2 types on `CallReplies`, the `RootCallReply` and the `InnerCallReply`. The `RootCallReply` has no life time restriction. The `InnerCallReply` can not outlive the `RootCallReply`  it was retrieved from.

To make the user life easier, both, the `RootCallReply` and the `InnerCallReply` implements the `CallReply` trait that allows to write a generic code that will work on both.