Allow the crate to operate when the Redis API isn't available.
Such situations occur when, for example, the crate which uses
the redismodule-rs crate as a dependency is run for testing, so
without Redis available. In this case, there is no allocator available
as well and it will lead to a panic. To enable both the situations to
work correctly under all circumstances, this change introduces a
fallback mechanism back to the system allocator (default Rust allocator)
which is always available.

Note that the changes are zero-cost: the code before this changes used
unwrap() which also had checks for the value being non-null and would
panic at runtime if it was null, the current code does this check as
well but instead of panicking reverts back to the system allocator.