Let the redis module implementation register the allocator.
This fixes lots of issues which were present due to a global allocator
being set forcefully by this crate. This is undesirable for the modules
dependent on this crate, as this becomes a trouble when running unit
tests, for example, and requires falling back to the system allocator.
As long as the Redis allocator is only required when running the library
as a redis module rather than always, setting it globally should be
avoided.