Don't run the example code when running tests.
The examples are not tests and shouldn't be run when the test are run.
Also, the examples are a redis module and can't be run without being
loaded by redis as a redis module and will always fail to launch without
redis, leading to the test errors due to the allocator being unavailable.
Also, the examples have no tests in those, so there is nothing to run.