Post notification API and API versioning (#304)
The PR is a proposal for API versioning using feature flags.
The PR introduce 4 new feature flags:
* `min-redis-compatibility-version-7-2`
* `min-redis-compatibility-version-7-0`
* `min-redis-compatibility-version-6-2`
* `min-redis-compatibility-version-6-0` - default

User should set one of those feature flag which will indicate the minimum Redis version he want to be compatible with. Base on this, redismodule-rs will decide which API can be used and will manipulate existing API so it can be used conditionally.

For example, take the following API:
```
pub fn add_post_notification_job<F: Fn(&Context)>(self, callback: F)
```

This API require Redis version 7.2, so if `min-redis-compatibility-version-7-2` feature is set the API is exposed as is. But if some other `min-redis-compatibility-version-*` is set, the API is changed to return `APIResult` which will allow to conditionally use this API if exists on the current Redis instance.

The PR also adds support for post notifications jobs and demonstrate how to use the new `min-redis-compatibility-version-*` flags.

Depends on: https://github.com/RedisLabsModules/redismodule-rs/pull/302