Provide the standard logging facility implementation.
As the crate is mostly used as a redis module, the logging will most
likely be only useful when printed on the server. As many rustaceans are
familiar and used to using the de-facto standard "log" crate, it makes
sense to provide the logging implementation for this crate to enable the
most-commonly used way of logging messages in Rust.