Revert the "LogLevel" type in the root.
To preserve the API compatibility with older versions, it is
necessary to allow to continue using the old "LogLevel" type.

This commit brings it back but not as a enum as it had been before it
changed, but as a type alias to the new enum. It also marks it as
deprecated with a hint.