fix: add `to_raw_parts` method.
This change introduces the `to_raw_parts` method as a counterpart to `from_raw_parts`. It also marks `from_raw_parts` as unsafe adding documentation to both.

Additionally this fixes an unsoundness bug in in `KeyCursor` that was leading to possible use-after free of the key name provided to the scan callback.